unholy levels of cleanup and optimization
