+++ index.html
@@ -2,8 +2,38 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/icon.png" />
+    <link rel="icon" type="image/svg+xml" href="/icon.webp" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <meta
+      name="description"
+      content="Noureldeen Ahmed's portfolio - Software Engineer specializing in React, TypeScript, and modern web development. Featuring interactive projects like Plant Together, Music Circle, and ML-driven content creation."
+    />
+    <meta
+      name="keywords"
+      content="software engineer, web developer, React, TypeScript, portfolio, Noureldeen Ahmed"
+    />
+    <meta name="author" content="Noureldeen Ahmed" />
+    <meta property="og:title" content="nnourr - Software Engineer Portfolio" />
+    <meta
+      property="og:description"
+      content="Explore my journey in software engineering through interactive projects and experiences. Featuring modern web development with React, TypeScript, and ML."
+    />
+    <meta property="og:type" content="website" />
+    <meta name="twitter:card" content="summary_large_image" />
+    <meta name="twitter:title" content="nnourr - Software Engineer Portfolio" />
+    <meta
+      name="twitter:description"
+      content="Modern web development portfolio featuring React, TypeScript, and innovative projects in software engineering."
+    />
+    <meta name="twitter:creator" content="@nnnoureldeennn" />
+    <meta
+      name="twitter:image"
+      content="https://nnourr.tech/headshotcropped.png"
+    />
+    <meta
+      name="twitter:image:alt"
+      content="Noureldeen Ahmed - Software Engineer"
+    />
     <title>nnourr</title>
     <link rel="preconnect" href="https://fonts.googleapis.com" />
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
