+++ src/components/Home/Experience.tsx
@@ -26,7 +26,7 @@ function Experience({
 }: ExperienceProps) {
   return (
     <div
-      className={`text-contrast md:hover:bg-secondary/40 flex gap-2 rounded-xl transition-all duration-300 ease-in-out md:gap-4 md:px-4 md:py-2 ${className}`}
+      className={`text-contrast md:hover:bg-secondary/40 group flex gap-2 rounded-xl transition-all duration-300 ease-in-out md:gap-4 md:px-4 md:py-2 ${className}`}
     >
       <div className="flex flex-col items-center gap-2">
         {/* Company Logo */}
@@ -73,7 +73,7 @@ function Experience({
           {blogLink && (
             <Link
               to={blogLink}
-              className="text-accent after:bg-accent relative cursor-pointer transition-all duration-150 ease-in-out after:absolute after:-inset-x-2 after:-inset-y-1 after:-z-10 after:rounded-full after:opacity-0 after:transition-all after:duration-150 after:ease-in-out hover:text-white hover:after:opacity-100"
+              className="md:text-contrast/90 hover:after:bg-accent md:group-hover:after:bg-secondary after:bg-secondary/40 text-accent md:group-hover:text-accent md:hover:after:bg-accent relative cursor-pointer transition-all duration-150 ease-in-out after:absolute after:-inset-x-1 after:-inset-y-0.5 after:-z-10 after:rounded-2xl after:transition-all after:duration-150 after:ease-in-out md:after:inset-0 md:after:bg-transparent md:group-hover:after:-inset-x-2 md:group-hover:after:-inset-y-1 md:group-hover:after:rounded-full md:hover:text-black"
             >
               read more
             </Link>
