+++ src/components/Home/Project.tsx
@@ -58,7 +58,11 @@ function Project({
         className={`flex flex-col gap-2 ${reverse ? 'md:flex-row-reverse' : 'md:flex-row'} md:gap-4`}
       >
         {/* Project Image */}
-        <div className="relative opacity-90 transition-all duration-300 ease-in-out active:scale-x-85 active:scale-y-95 md:w-1/2">
+        <a
+          href={links[0].url}
+          target="_blank"
+          className="after:bg-secondary/70 after:text-contrast/100 after:decoration-accent relative opacity-90 transition-all duration-300 ease-in-out after:absolute after:inset-0 after:z-10 after:flex after:items-center after:justify-center after:text-center after:font-sans after:text-xl after:font-bold after:underline after:decoration-2 after:opacity-0 after:transition-all after:duration-300 after:ease-in-out after:content-['explore!'] hover:after:opacity-100 md:w-1/2"
+        >
           <img
             src={imageSrc}
             alt={imageAlt}
@@ -69,7 +73,7 @@ function Project({
             alt={imageAlt + ' blurred'}
             className="absolute top-1/2 right-0 -z-10 h-24 w-full -translate-y-1/2 rounded-lg object-cover opacity-90 blur-lg contrast-90 saturate-500 md:right-6 md:h-32 md:w-11/12 md:opacity-55"
           />
-        </div>
+        </a>
         <div className="flex flex-col gap-2 md:w-1/2">
           {/* Description */}
           <div className="font-sans text-sm leading-relaxed">{children}</div>
@@ -90,6 +94,7 @@ function Project({
                 size="sm"
                 className="flex items-center gap-2"
                 onClick={() => window.open(link.url, '_blank')}
+                title={`Open ${title} ${link.label} in new tab`}
               >
                 <FontAwesomeIcon icon={link.icon} />
                 {link.label}
