+++ src/pages/Blog/rbc.tsx
@@ -9,18 +9,18 @@ function RbcBlog() {
 
   useEffect(() => {
     setOpen();
-  }, []);
+  }, [setOpen]);
 
   return (
     <div className="relative mt-6 pt-10 md:pt-22">
       <div className="fade-in animate-in 32 absolute -top-24 -right-4 -left-4 -z-50 h-32 duration-1000 ease-in-out md:h-52">
         <img
-          src="/rbc-banner.jpeg"
+          src="/rbc-banner.webp"
           alt="rbc banner"
           className="h-full w-full rounded-2xl object-cover object-right opacity-90"
         />
         <img
-          src="/rbc-banner.jpeg"
+          src="/rbc-banner.webp"
           alt="rbc banner"
           className="absolute top-0 left-0 -z-10 h-2/3 w-full rounded-2xl object-cover blur-2xl md:opacity-70"
         />
